requirements.txt
discord.py==1.3.4
gspread==3.6.0
pandas==1.0.4
python-dotenv==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
cffi==1.14.0
websockets==8.1
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
multidict==4.7.6
chardet==3.0.4
yarl==1.4.2
idna==2.9